@@ -20,29 +20,29 @@
     "test:e2e": "export NODE_ENV=test-e2e && jest --config ./tests/jest-e2e.json --detectOpenHandles --forceExit"
   },
   "dependencies": {
-    "@nestjs/common": "^6.2.4",
-    "@nestjs/core": "^6.2.4",
+    "@nestjs/common": "^6.3.1",
+    "@nestjs/core": "^6.3.1",
     "@nestjs/cqrs": "^6.0.0",
-    "@nestjs/microservices": "^6.2.4",
+    "@nestjs/microservices": "^6.3.1",
     "@nestjs/mongoose": "^6.1.2",
-    "@nestjs/platform-express": "^6.2.4",
+    "@nestjs/platform-express": "^6.3.1",
     "@nestjs/swagger": "^3.0.2",
-    "@nestjs/websockets": "^6.2.4",
-    "bull": "^3.9.1",
+    "@nestjs/websockets": "^6.3.1",
+    "bull": "^3.10.0",
     "class-transformer": "^0.2.3",
     "class-validator": "^0.9.1",
     "connect-mongo": "^2.0.3",
     "cookie-parser": "^1.4.4",
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
-    "express-rate-limit": "^4.0.2",
+    "express-rate-limit": "^5.0.0",
     "express-session": "^1.16.1",
     "hbs": "^4.0.4",
     "helmet": "^3.18.0",
     "i18n": "^0.8.3",
     "joi": "^14.3.1",
-    "mongodb": "^3.2.6",
-    "mongoose": "^5.5.11",
+    "mongodb": "^3.2.7",
+    "mongoose": "^5.5.14",
     "node-jose": "^1.1.3",
     "oauth2orize": "^1.11.0",
     "oauth2orize-openid": "^0.4.0",
@@ -57,32 +57,32 @@
     "reflect-metadata": "^0.1.13",
     "rxjs": "^6.5.2",
     "speakeasy": "^2.0.0",
-    "swagger-ui-express": "^4.0.5",
+    "swagger-ui-express": "^4.0.6",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
-    "@nestjs/testing": "^6.2.4",
+    "@nestjs/testing": "^6.3.1",
     "@types/bull": "^3.5.14",
     "@types/connect-mongo": "0.0.42",
-    "@types/express": "^4.16.1",
+    "@types/express": "^4.17.0",
     "@types/jest": "^24.0.13",
     "@types/joi": "^14.3.3",
-    "@types/mongoose": "^5.5.3",
-    "@types/node": "^12.0.3",
+    "@types/mongoose": "^5.5.5",
+    "@types/node": "^12.0.7",
     "@types/oauth": "^0.9.1",
     "@types/passport-local": "^1.0.33",
     "@types/supertest": "^2.0.7",
     "jest": "^24.8.0",
     "nodemon": "^1.19.1",
-    "prettier": "^1.17.1",
+    "prettier": "^1.18.2",
     "prettier-check": "^2.0.0",
     "supertest": "^4.0.2",
     "ts-jest": "^24.0.2",
-    "ts-loader": "^6.0.1",
+    "ts-loader": "^6.0.2",
     "ts-node": "^8.2.0",
     "tsconfig-paths": "^3.8.0",
-    "tslint": "^5.16.0",
-    "typescript": "^3.4.5"
+    "tslint": "^5.17.0",
+    "typescript": "^3.5.1"
   },
   "jest": {
     "testEnvironment": "node",
