@@ -19,12 +19,12 @@
   },
   "private": true,
   "dependencies": {
-    "@nestjs/common": "^6.2.4",
-    "@nestjs/core": "^6.2.4",
+    "@nestjs/common": "^6.3.1",
+    "@nestjs/core": "^6.3.1",
     "@nestjs/cqrs": "^6.0.0",
-    "@nestjs/platform-express": "^6.2.4",
+    "@nestjs/platform-express": "^6.3.1",
     "@nestjs/swagger": "^3.0.2",
-    "@nestjs/typeorm": "^6.1.1",
+    "@nestjs/typeorm": "^6.1.2",
     "class-transformer": "^0.2.3",
     "class-validator": "^0.9.1",
     "cron": "^1.7.1",
@@ -32,29 +32,29 @@
     "fastify-formbody": "^3.1.0",
     "form-data": "^2.3.3",
     "joi": "^14.3.1",
-    "mongodb": "^3.2.6",
+    "mongodb": "^3.2.7",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^6.5.2",
-    "swagger-ui-express": "^4.0.5",
-    "typeorm": "^0.2.17"
+    "swagger-ui-express": "^4.0.6",
+    "typeorm": "^0.2.18"
   },
   "devDependencies": {
-    "@nestjs/testing": "^6.2.4",
-    "@types/express": "^4.16.1",
+    "@nestjs/testing": "^6.3.1",
+    "@types/express": "^4.17.0",
     "@types/jest": "^24.0.13",
-    "@types/node": "^12.0.3",
+    "@types/node": "^12.0.7",
     "@types/supertest": "^2.0.7",
     "jest": "^24.8.0",
     "nodemon": "^1.19.1",
-    "prettier": "^1.17.1",
+    "prettier": "^1.18.2",
     "prettier-check": "^2.0.0",
     "supertest": "^4.0.2",
     "ts-jest": "^24.0.2",
-    "ts-loader": "^6.0.1",
+    "ts-loader": "^6.0.2",
     "ts-node": "^8.2.0",
     "tsconfig-paths": "^3.8.0",
-    "tslint": "^5.16.0",
-    "typescript": "^3.4.5"
+    "tslint": "^5.17.0",
+    "typescript": "^3.5.1"
   },
   "jest": {
     "moduleDirectories": [
