@@ -1,18 +1,16 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { Client, CLIENT } from './client/client.schema';
 import { ClientService } from './client/client.service';
-import { SCOPE, Scope } from './scope/scope.schema';
 import { ScopeService } from './scope/scope.service';
+import { ClientManagementModuleEntities } from './entities';
 
 @Module({
-  imports: [
-    MongooseModule.forFeature([
-      { name: CLIENT, schema: Client },
-      { name: SCOPE, schema: Scope },
-    ]),
-  ],
+  imports: [MongooseModule.forFeature(ClientManagementModuleEntities)],
   providers: [ClientService, ScopeService],
-  exports: [ClientService, ScopeService],
+  exports: [
+    ClientService,
+    ScopeService,
+    MongooseModule.forFeature(ClientManagementModuleEntities),
+  ],
 })
 export class ClientManagementEntitiesModule {}
