@@ -1,19 +1,14 @@
 import { Module, HttpModule } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import {
-  SERVER_SETTINGS,
-  ServerSettings,
-} from './server-settings/server-settings.schema';
 import { ServerSettingsService } from './server-settings/server-settings.service';
+import { SystemSettingsModuleEntites } from './entities';
 
 @Module({
-  imports: [
-    MongooseModule.forFeature([
-      { name: SERVER_SETTINGS, schema: ServerSettings },
-    ]),
-    HttpModule,
-  ],
+  imports: [MongooseModule.forFeature(SystemSettingsModuleEntites), HttpModule],
   providers: [ServerSettingsService],
-  exports: [ServerSettingsService],
+  exports: [
+    ServerSettingsService,
+    MongooseModule.forFeature(SystemSettingsModuleEntites),
+  ],
 })
 export class SystemSettingsEntitiesModule {}
