@@ -1,4 +1,4 @@
-import { Module, Global } from '@nestjs/common';
+import { Module, Global, HttpModule } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Oauth2ProviderController } from './controllers/oauth2-provider/oauth2-provider.controller';
 import { OAuth2Provider } from './entities/oauth2-provider/oauth2-provider.entity';
@@ -8,10 +8,15 @@ import { OAuth2ClientCommandHandlers } from './commands';
 import { OAuth2ClientQueryHandlers } from './queries';
 import { OAuth2ClientEventHandlers } from './events';
 import { OAuth2Token } from './entities/oauth2-token/oauth2-token.entity';
+import { CqrsModule } from '@nestjs/cqrs';
 
 Global();
 @Module({
-  imports: [TypeOrmModule.forFeature([OAuth2Provider, OAuth2Token])],
+  imports: [
+    TypeOrmModule.forFeature([OAuth2Provider, OAuth2Token]),
+    CqrsModule,
+    HttpModule,
+  ],
   providers: [
     ...OAuth2ClientEntityProviders,
     ...OAuth2ClientAggregates,
