@@ -8,6 +8,7 @@ import {
   UNDO,
 } from '../../constants/messages';
 import { MatSnackBar } from '@angular/material';
+import { DURATION, UNDO_DURATION } from '../../constants/common';
 
 @Component({
   selector: 'app-auth-settings',
@@ -102,7 +103,7 @@ export class AuthSettingsComponent implements OnInit {
       )
       .subscribe({
         next: response => {
-          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: DURATION });
         },
         error: error => {},
       });
@@ -117,7 +118,9 @@ export class AuthSettingsComponent implements OnInit {
         )
         .subscribe({
           next: success => {
-            this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+            this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, {
+              duration: DURATION,
+            });
           },
           error: error => {},
         });
@@ -128,7 +131,9 @@ export class AuthSettingsComponent implements OnInit {
     this.flagDeleteUserSessions = true;
     this.disableDeleteSessions = true;
     this.disableDeleteTokens = true;
-    const snackBar = this.snackBar.open(DELETING, UNDO, { duration: 10000 });
+    const snackBar = this.snackBar.open(DELETING, UNDO, {
+      duration: UNDO_DURATION,
+    });
 
     snackBar.afterDismissed().subscribe({
       next: dismissed => {
@@ -158,7 +163,9 @@ export class AuthSettingsComponent implements OnInit {
     this.flagDeleteUserSessions = true;
     this.disableDeleteSessions = true;
     this.disableDeleteTokens = true;
-    const snackBar = this.snackBar.open(DELETING, UNDO, { duration: 10000 });
+    const snackBar = this.snackBar.open(DELETING, UNDO, {
+      duration: UNDO_DURATION,
+    });
 
     snackBar.afterDismissed().subscribe({
       next: dismissed => {
