@@ -10,7 +10,7 @@ import {
   CREATE_ERROR,
   UPDATE_ERROR,
 } from '../../constants/messages';
-import { NEW_ID } from '../../constants/common';
+import { NEW_ID, DURATION } from '../../constants/common';
 
 export const ROLE_LIST_ROUTE = '/role/list';
 
@@ -52,11 +52,11 @@ export class RoleComponent implements OnInit {
       .createRole(this.roleForm.controls.roleName.value)
       .subscribe({
         next: success => {
-          this.snackBar.open(CREATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+          this.snackBar.open(CREATE_SUCCESSFUL, CLOSE, { duration: DURATION });
           this.router.navigateByUrl(ROLE_LIST_ROUTE);
         },
         error: error =>
-          this.snackBar.open(CREATE_ERROR, CLOSE, { duration: 2000 }),
+          this.snackBar.open(CREATE_ERROR, CLOSE, { duration: DURATION }),
       });
   }
 
@@ -65,11 +65,11 @@ export class RoleComponent implements OnInit {
       .updateRole(this.uuid, this.roleForm.controls.roleName.value)
       .subscribe({
         next: success => {
-          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: DURATION });
           this.router.navigateByUrl(ROLE_LIST_ROUTE);
         },
         error: error =>
-          this.snackBar.open(UPDATE_ERROR, CLOSE, { duration: 2000 }),
+          this.snackBar.open(UPDATE_ERROR, CLOSE, { duration: DURATION }),
       });
   }
 }
