@@ -3,7 +3,7 @@ import { FormArray, FormGroup, FormControl } from '@angular/forms';
 import { SocialLoginService } from './social-login.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material';
-import { NEW_ID } from '../../constants/common';
+import { NEW_ID, DURATION } from '../../constants/common';
 import {
   SOCIAL_LOGIN_ERROR,
   SOCIAL_LOGIN_UPDATED,
@@ -152,11 +152,13 @@ export class SocialLoginComponent implements OnInit {
       )
       .subscribe({
         next: success => {
-          this.snackBar.open(SOCIAL_LOGIN_CREATED, CLOSE, { duration: 2000 });
+          this.snackBar.open(SOCIAL_LOGIN_CREATED, CLOSE, {
+            duration: DURATION,
+          });
           this.router.navigateByUrl(SOCIAL_LOGIN_LIST_ROUTE);
         },
         error: error =>
-          this.snackBar.open(SOCIAL_LOGIN_ERROR, CLOSE, { duration: 2000 }),
+          this.snackBar.open(SOCIAL_LOGIN_ERROR, CLOSE, { duration: DURATION }),
       });
   }
 
@@ -188,11 +190,13 @@ export class SocialLoginComponent implements OnInit {
       )
       .subscribe({
         next: success => {
-          this.snackBar.open(SOCIAL_LOGIN_UPDATED, CLOSE, { duration: 2000 });
+          this.snackBar.open(SOCIAL_LOGIN_UPDATED, CLOSE, {
+            duration: DURATION,
+          });
           this.router.navigateByUrl(SOCIAL_LOGIN_LIST_ROUTE);
         },
         error: error =>
-          this.snackBar.open(UPDATE_ERROR, CLOSE, { duration: 2000 }),
+          this.snackBar.open(UPDATE_ERROR, CLOSE, { duration: DURATION }),
       });
   }
 }
