@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { CommunicationSettingsService } from './communication-settings.service';
 import { MatSnackBar } from '@angular/material';
 import { FormGroup, FormControl } from '@angular/forms';
-import { UPDATE_SUCCESSFUL, CLOSE } from 'src/app/constants/messages';
+import { UPDATE_SUCCESSFUL, CLOSE } from '../../constants/messages';
+import { DURATION } from '../../constants/common';
 
 export interface EmailAccount {
   host: string;
@@ -84,7 +85,7 @@ export class CommunicationSettingsComponent implements OnInit {
       )
       .subscribe({
         next: response => {
-          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: DURATION });
         },
         error: error => {},
       });
