@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl } from '@angular/forms';
 import { IdpSettingsService } from './idp-settings.service';
 import { MatSnackBar } from '@angular/material';
-import { UPDATE_SUCCESSFUL, CLOSE } from 'src/app/constants/messages';
+import { UPDATE_SUCCESSFUL, CLOSE } from '../../constants/messages';
+import { DURATION } from '../../constants/common';
 
 export interface CloudStorage {
   name: string;
@@ -71,7 +72,7 @@ export class IdpSettingsComponent implements OnInit {
       )
       .subscribe({
         next: response => {
-          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: DURATION });
         },
         error: error => {},
       });
